backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.32
alembic==1.13.2
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.4.0
python-multipart==0.0.9
redis==5.0.8
itsdangerous==2.2.0
starlette==0.37.2
httpx==0.27.0
